# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CorridorOps",
    "description": "CorridorOps - AI-Powered Automatic Block Planning System for Indian Railways (SIH PS #26027)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ingest/track-sections": {
      "post": {
        "tags": [
          "Ingestion & Infrastructure"
        ],
        "summary": "Ingest Track Sections",
        "operationId": "ingest_track_sections_api_v1_ingest_track_sections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeoJSONFeatureCollection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/track-sections": {
      "get": {
        "tags": [
          "Ingestion & Infrastructure"
        ],
        "summary": "List Track Sections",
        "operationId": "list_track_sections_api_v1_track_sections_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "division",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Division"
            }
          },
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_TrackSectionSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/track-sections/{section_id}": {
      "get": {
        "tags": [
          "Ingestion & Infrastructure"
        ],
        "summary": "Get Track Section Detail",
        "operationId": "get_track_section_detail_api_v1_track_sections__section_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "section_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Section Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_TrackSectionDetailSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest/train-schedule": {
      "post": {
        "tags": [
          "Ingestion & Infrastructure"
        ],
        "summary": "Ingest Train Schedule",
        "operationId": "ingest_train_schedule_api_v1_ingest_train_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/TrainScheduleCreate"
                },
                "type": "array",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/train-schedule": {
      "get": {
        "tags": [
          "Ingestion & Infrastructure"
        ],
        "summary": "List Train Schedules",
        "operationId": "list_train_schedules_api_v1_train_schedule_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "section_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse_List_TrainScheduleSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeoJSONFeature": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "Feature"
          },
          "properties": {
            "$ref": "#/components/schemas/GeoJSONFeatureProperties"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeoJSONFeatureGeometry"
          }
        },
        "type": "object",
        "required": [
          "properties",
          "geometry"
        ],
        "title": "GeoJSONFeature"
      },
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/GeoJSONFeature"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "GeoJSONFeatureCollection"
      },
      "GeoJSONFeatureGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "coordinates": {
            "title": "Coordinates"
          }
        },
        "type": "object",
        "required": [
          "type",
          "coordinates"
        ],
        "title": "GeoJSONFeatureGeometry"
      },
      "GeoJSONFeatureProperties": {
        "properties": {
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone",
            "default": "NR"
          },
          "division": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Division",
            "default": "DLI"
          },
          "start_station": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Station",
            "default": "START"
          },
          "end_station": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Station",
            "default": "END"
          }
        },
        "type": "object",
        "required": [
          "section_id"
        ],
        "title": "GeoJSONFeatureProperties"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StandardResponse_List_TrackSectionSchema__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TrackSectionSchema"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[List[TrackSectionSchema]]"
      },
      "StandardResponse_List_TrainScheduleSchema__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TrainScheduleSchema"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[List[TrainScheduleSchema]]"
      },
...
      "StandardResponse_TrackSectionDetailSchema_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TrackSectionDetailSchema"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[TrackSectionDetailSchema]"
      },
...
      "StandardResponse_dict_": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StandardResponse[dict]"
      },
...
      "TrackSectionDetailSchema": {
        "properties": {
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "geometry": {
            "type": "string",
            "title": "Geometry"
          },
          "zone": {
            "type": "string",
            "title": "Zone"
          },
          "division": {
            "type": "string",
            "title": "Division"
          },
          "start_station": {
            "type": "string",
            "title": "Start Station"
          },
          "end_station": {
            "type": "string",
            "title": "End Station"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "active_events": {
            "items": {},
            "type": "array",
            "title": "Active Events",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "section_id",
          "geometry",
          "zone",
          "division",
          "start_station",
          "end_station",
          "id"
        ],
        "title": "TrackSectionDetailSchema"
      },
      "TrackSectionSchema": {
        "properties": {
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "geometry": {
            "type": "string",
            "title": "Geometry"
          },
          "zone": {
            "type": "string",
            "title": "Zone"
          },
          "division": {
            "type": "string",
            "title": "Division"
          },
          "start_station": {
            "type": "string",
            "title": "Start Station"
          },
          "end_station": {
            "type": "string",
            "title": "End Station"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "section_id",
          "geometry",
          "zone",
          "division",
          "start_station",
          "end_station",
          "id"
        ],
        "title": "TrackSectionSchema"
      },
      "TrainScheduleCreate": {
        "properties": {
          "train_number": {
            "type": "string",
            "title": "Train Number"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "arrival_time": {
            "type": "string",
            "title": "Arrival Time"
          },
          "departure_time": {
            "type": "string",
            "title": "Departure Time"
          },
          "days_of_week": {
            "type": "string",
            "title": "Days Of Week"
          },
          "train_type": {
            "type": "string",
            "title": "Train Type"
          }
        },
        "type": "object",
        "required": [
          "train_number",
          "section_id",
          "arrival_time",
          "departure_time",
          "days_of_week",
          "train_type"
        ],
        "title": "TrainScheduleCreate"
      },
      "TrainScheduleSchema": {
        "properties": {
          "train_number": {
            "type": "string",
            "title": "Train Number"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "arrival_time": {
            "type": "string",
            "title": "Arrival Time"
          },
          "departure_time": {
            "type": "string",
            "title": "Departure Time"
          },
          "days_of_week": {
            "type": "string",
            "title": "Days Of Week"
          },
          "train_type": {
            "type": "string",
            "title": "Train Type"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "train_number",
          "section_id",
          "arrival_time",
          "departure_time",
          "days_of_week",
          "train_type",
          "id"
        ],
        "title": "TrainScheduleSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}